requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
h5py>=3.7.0
pytest>=7.0.0
openctp_ctp>=0.5.0
fire>=0.4.0
